Partitioned test cases checkers and initializing of test case project environment.
